All separated. Working on setting up local dev
