--- Clipboard.html
@@ -1,95 +1,0 @@
-<table>
-  <thead>
-    <tr>
-      <th colspan="1"></th>
-      <th colspan="6">Story Point Counts</th>
-      <th colspan="2"></th>
-    </tr>
-    <tr>
-      <th>Issue</th>
-      <th>Total</th>
-      <th>H.J.</th>
-      <th>E.A.</th>
-      <th>C.H.</th>
-      <th>J.S.</th>
-      <th>% Complete</th>
-      <th>Notes</th>
-    </tr>
-  </thead>
-
-  <tbody>
-    <tr>
-      <td>
-        <a href=https://github.com/wsu-cpts489-fa21/tp-good-team/issues/124>#24</a></td>
-      <td>8</td>
-      <td>0</td>
-      <td>0</td>
-      <td>0</td>
-      <td>8</td>
-      <td>100</td>
-      <td>
-          Created DB schema. Initially structured as a table, but later switched UI to look more like a Feed Post.
-      </td>
-      </tr>
-    </tr>
-    <tr>
-      <td><a href="https://github.com/wsu-cpts489-fa21/tp-good-team/issues/120">#120</a></td>
-      <td>3</td>
-      <td>3</td>
-      <td>0</td>
-      <td>0</td>
-      <td>0</td>
-      <td>100</td>
-      <td>
-        Added to the schema to take count of the number received. Searched online for free svg icons to use.
-        Implemented this by giving specific rows onClick() abilities.
-      </td>
-    </tr>
-    <tr>
-        <td><a href="https://github.com/wsu-cpts489-fa21/tp-good-team/issues/123">#123</a></td>
-        <td>3</td>
-        <td>3</td>
-        <td>0</td>
-        <td>0</td>
-        <td>0</td>
-        <td>100</td>
-        <td>
-          Accomplished by creating a `Comment` icon that users can click. Creating document-wide handlers affected handlers that existed in the children.
-        </td>
-      </tr>
-      <tr>
-        <td><a href="https://github.com/wsu-cpts489-fa21/tp-good-team/issues/121">#121</a></td>
-        <td>3</td>
-        <td>0</td>
-        <td>3</td>
-        <td>0</td>
-        <td>0</td>
-        <td>100</td>
-        <td>
-          While rendering the table of Posts, we can save the index and use it to link to specific handlers
-        </td>
-      </tr>
-      <tr>
-        <td><a href="https://github.com/wsu-cpts489-fa21/tp-good-team/issues/122">#122</a></td>
-        <td>3</td>
-        <td>0</td>
-        <td>3</td>
-        <td>0</td>
-        <td>0</td>
-        <td>100</td>
-        <td>
-          Updated `RoundForm` to include a box to check. If this is checked we don't render this Post
-        </td>
-      </tr>
-      
-    <tr>
-      <td colspan="2" align="right"><b>Totals:</b></td>
-      <td><b>8</b></td>
-      <td><b>3</b></td>
-      <td><b>1</b></td>
-      <td><b>1</b></td>
-      <td><b>0.8</b></td>
-      <td colspan="2"></td>
-    </tr>
-  </tbody>
-</table>
