+++ client/src/components/LoginPage.js
@@ -242,10 +242,10 @@ class LoginPage extends React.Component {
             className="btn btn-primary fm-primary-btn"
           >
             test
-            <FontAwesomeIcon
+            {/* <FontAwesomeIcon
               icon={this.state.loginBtnIcon}
               className={this.state.loginBtnIcon === "spinner" ? "fa-spin" : ""}
-            />
+            /> */}
             &nbsp;{this.state.loginBtnLabel}
           </button>
         </form>
