--- package.json
@@ -1,60 +1,0 @@
-{
-  "name": "speedscore-book",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "dev": "nodemon --exec node -r dotenv/config server.js",
-    "start": "node -r dotenv/config server.js",
-    "test": "npx testcafe chrome tests/*.js",
-    "b": "cd client && npm run build && cd .. && npm run dev",
-    "install-all-deps": "npm install -g testcafe && npm i testcafe --save-dev && npm i nanoid npm install passport-google-oauth20 && npm install passport-local && npm install bcrypt && npm install mongoose && npm install passport-local && npm install passport && npm install passport-github && npm install dotenv && cd client && npm install && npm install --save @fortawesome/free-brands-svg-icons && npm install bootstrap-social && npm run build && cd .. && npm i"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/FSWebDevBook/speedscore-book.git"
-  },
-  "keywords": [
-    "heroku",
-    "react",
-    "express",
-    "node"
-  ],
-  "author": "C. Hundhausen",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/FSWebDevBook/speedscore-book/issues"
-  },
-  "homepage": "https://github.com/FSWebDevBook/speedscore-book#readme",
-  "dependencies": {
-    "bcrypt": "^5.1.0",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-session": "^1.17.2",
-    "install": "^0.13.0",
-    "mongoose": "^6.9.1",
-    "npm": "^8.19.4",
-    "passport": "^0.5.3",
-    "passport-github": "^1.1.0",
-    "passport-google-oauth20": "^2.0.0",
-    "passport-local": "^1.0.0",
-    "path": "^0.12.7",
-    "react-bootstrap": "^2.0.3",
-    "react-modal": "^3.14.4",
-    "url": "^0.11.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.8",
-    "@babel/node": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@testing-library/react": "^12.1.2",
-    "nanoid": "^3.3.4",
-    "nodemon": "^2.0.14",
-    "testcafe": "^1.20.1"
-  },
-  "engines": {
-    "node": "14.x"
-  },
-  "type": "module"
-}
