--- run.sh
@@ -1,4 +1,0 @@
-cd client
-npm run build
-cd ..
-npm run dev
